backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-dotenv==1.0.1
solana==0.35.0
solders==0.21.0
base58==2.1.1
httpx==0.27.0
PyJWT[crypto]==2.9.0
